+++ src/Action/Identity/IdentityOverview.php
@@ -1,0 +1,195 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Heptacom\HeptaConnect\Storage\ShopwareDal\Action\Identity;
+
+use Doctrine\DBAL\Connection;
+use Doctrine\DBAL\Driver\ResultStatement;
+use Doctrine\DBAL\FetchMode;
+use Doctrine\DBAL\Query\QueryBuilder;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Overview\IdentityOverviewCriteria;
+use Heptacom\HeptaConnect\Storage\Base\Action\Identity\Overview\IdentityOverviewResult;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Identity\IdentityOverviewActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Exception\InvalidOverviewCriteriaException;
+use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\MappingNodeStorageKey;
+use Heptacom\HeptaConnect\Storage\ShopwareDal\StorageKey\PortalNodeStorageKey;
+use Shopware\Core\Defaults;
+use Shopware\Core\Framework\Uuid\Uuid;
+
+class IdentityOverview implements IdentityOverviewActionInterface
+{
+    private ?QueryBuilder $builder = null;
+
+    private Connection $connection;
+
+    public function __construct(Connection $connection)
+    {
+        $this->connection = $connection;
+    }
+
+    public function overview(IdentityOverviewCriteria $criteria): iterable
+    {
+        $builder = $this->getBuilderCached();
+        $mappingNodeKeyFilter = $criteria->getMappingNodeKeyFilter();
+        $entityTypeFilter = $criteria->getEntityTypeFilter();
+        $externalIdFilter = $criteria->getExternalIdFilter();
+        $portalNodeKeyFilter = $criteria->getPortalNodeKeyFilter();
+
+        if ($mappingNodeKeyFilter->count() > 0) {
+            $mappingNodeIds = [];
+
+            foreach ($mappingNodeKeyFilter as $mappingNodeKey) {
+                if (!$mappingNodeKey instanceof MappingNodeStorageKey) {
+                    throw new InvalidOverviewCriteriaException($criteria, 1643877525, new UnsupportedStorageKeyException(\get_class($mappingNodeKey)));
+                }
+
+                $mappingNodeIds[] = \hex2bin($mappingNodeKey->getUuid());
+            }
+
+            $builder->andWhere($builder->expr()->in('mapping_node.id', ':mappingNodeIds'));
+            $builder->setParameter('mappingNodeIds', $mappingNodeIds, Connection::PARAM_STR_ARRAY);
+        }
+
+        if ($entityTypeFilter !== []) {
+            $builder->andWhere($builder->expr()->in('entity_type.type', ':entityTypes'));
+            $builder->setParameter('entityTypes', $entityTypeFilter, Connection::PARAM_STR_ARRAY);
+        }
+
+        if ($externalIdFilter !== []) {
+            $builder->andWhere($builder->expr()->in('mapping.external_id', ':externalIds'));
+            $builder->setParameter('externalIds', $externalIdFilter, Connection::PARAM_STR_ARRAY);
+        }
+
+        if ($portalNodeKeyFilter->count() > 0) {
+            $portalNodeIds = [];
+
+            foreach ($portalNodeKeyFilter as $portalNodeKey) {
+                if (!$portalNodeKey instanceof PortalNodeStorageKey) {
+                    throw new InvalidOverviewCriteriaException($criteria, 1643877526, new UnsupportedStorageKeyException(\get_class($portalNodeKey)));
+                }
+
+                $portalNodeIds[] = \hex2bin($portalNodeKey->getUuid());
+            }
+
+            $builder->andWhere($builder->expr()->in('portal_node.id', ':portalNodeIds'));
+            $builder->setParameter('portalNodeIds', $portalNodeIds, Connection::PARAM_STR_ARRAY);
+        }
+
+        foreach ($criteria->getSort() as $field => $direction) {
+            $dbalDirection = $direction === IdentityOverviewCriteria::SORT_ASC ? 'ASC' : 'DESC';
+            $dbalFieldName = null;
+
+            switch ($field) {
+                case IdentityOverviewCriteria::FIELD_CREATED:
+                    $dbalFieldName = 'mapping.created_at';
+
+                    break;
+                case IdentityOverviewCriteria::FIELD_ENTITY_TYPE:
+                    $dbalFieldName = 'entity_type.type';
+
+                    break;
+                case IdentityOverviewCriteria::FIELD_EXTERNAL_ID:
+                    $dbalFieldName = 'mapping.external_id';
+
+                    break;
+                case IdentityOverviewCriteria::FIELD_MAPPING_NODE:
+                    $dbalFieldName = 'mapping_node.id';
+
+                    break;
+                case IdentityOverviewCriteria::FIELD_PORTAL_NODE:
+                    $dbalFieldName = 'portal_node.id';
+
+                    break;
+            }
+
+            if ($dbalFieldName === null) {
+                throw new InvalidOverviewCriteriaException($criteria, 1643877527);
+            }
+
+            $builder->addOrderBy($dbalFieldName, $dbalDirection);
+        }
+
+        $builder->addOrderBy('mapping.id', 'ASC');
+
+        $pageSize = $criteria->getPageSize();
+
+        if ($pageSize !== null && $pageSize > 0) {
+            $page = $criteria->getPage();
+
+            $builder->setMaxResults($pageSize);
+
+            if ($page > 0) {
+                $builder->setFirstResult($page * $pageSize);
+            }
+        }
+
+        $statement = $builder->execute();
+
+        if (!$statement instanceof ResultStatement) {
+            throw new \LogicException('$builder->execute() should have returned a ResultStatement', 1643877528);
+        }
+
+        yield from \iterable_map(
+            $statement->fetchAll(FetchMode::ASSOCIATIVE),
+            static fn (array $row): IdentityOverviewResult => new IdentityOverviewResult(
+                new PortalNodeStorageKey(Uuid::fromBytesToHex((string) $row['portal_node_id'])),
+                new MappingNodeStorageKey(Uuid::fromBytesToHex((string) $row['mapping_node_id'])),
+                (string) $row['mapping_external_id'],
+                (string) $row['entity_type_type'],
+                /* @phpstan-ignore-next-line */
+                \date_create_immutable_from_format(Defaults::STORAGE_DATE_TIME_FORMAT, (string) $row['created_at']),
+            )
+        );
+    }
+
+    protected function getBuilderCached(): QueryBuilder
+    {
+        if (!$this->builder instanceof QueryBuilder) {
+            $this->builder = $this->getBuilder();
+            $this->builder->setFirstResult(0);
+            $this->builder->setMaxResults(null);
+            $this->builder->getSQL();
+        }
+
+        return clone $this->builder;
+    }
+
+    protected function getBuilder(): QueryBuilder
+    {
+        $builder = new QueryBuilder($this->connection);
+
+        $builder->from('heptaconnect_mapping', 'mapping')
+            ->innerJoin(
+                'mapping',
+                'heptaconnect_portal_node',
+                'portal_node',
+                $builder->expr()->eq('mapping.portal_node_id', 'portal_node.id')
+            )
+            ->innerJoin(
+                'mapping',
+                'heptaconnect_mapping_node',
+                'mapping_node',
+                $builder->expr()->eq('mapping.mapping_node_id', 'mapping_node.id')
+            )
+            ->innerJoin(
+                'mapping_node',
+                'heptaconnect_entity_type',
+                'entity_type',
+                $builder->expr()->eq('mapping_node.type_id', 'entity_type.id')
+            )
+            ->select([
+                'portal_node.id portal_node_id',
+                'mapping_node.id mapping_node_id',
+                'mapping.external_id mapping_external_id',
+                'entity_type.type entity_type_type',
+                'mapping.created_at created_at',
+            ])
+            ->andWhere($builder->expr()->isNull('portal_node.deleted_at'))
+            ->andWhere($builder->expr()->isNull('mapping_node.deleted_at'))
+            ->andWhere($builder->expr()->isNull('mapping.deleted_at'));
+
+        return $builder;
+    }
+}